Working on warnings produced by -Wall -Wextra. In particular, working on unused parameters. We have also changed includes - changed "petsc.h" to <petsc.h>
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@863 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ pde/src/common/Point.hpp
@@ -45,7 +45,7 @@ public:
 
     double operator[] (unsigned i) const
     {
-        assert(i<DIM); 
+        assert((int)i<DIM); 
         return mLocation[i];
     }
